Make an abstract class for SyncService common methods
* We want to re-use SyncServiceUtils background sync methods
